// src/services/models/exam.d.ts
import { Appointment } from "./appointment";

export type Exam = {
  appointment: Appointment;
  appointmentId: 1;
  createdAt: string;
  createdBy: string;
  description: string;
  isActive: string;
  modifiedAt: string;
  modifiedBy: string;
  testDate: string;
  testId: number;
  title: string;
};